parkinsons-speech-therapy/src/models/model_utils.py
import joblib
import numpy as np

def load_model(model_path):
    """Load the trained model from the specified path."""
    model = joblib.load(model_path)
    return model

def prepare_input(audio_data):
    """Preprocess the audio data for model input."""
    # Example preprocessing steps
    features = extract_features(audio_data)
    normalized_features = normalize_data(features)
    return np.array(normalized_features).reshape(1, -1)

def extract_features(audio_data):
    """Extract relevant features from the audio input."""
    # Placeholder for feature extraction logic
    return []

def normalize_data(features):
    """Normalize the audio features."""
    # Placeholder for normalization logic
    return features