parkinsons-speech-therapy/requirements.txt
streamlit>=1.32.0
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.45.0
tokenizers>=0.20.0,<0.21.0
peft>=0.7.0
bitsandbytes>=0.41.0
accelerate>=0.21.0
llama-cpp-python>=0.2.0
numpy>=1.24.0
librosa>=0.10.0
sounddevice>=0.4.6
scipy>=1.10.0
matplotlib>=3.7.0
pandas>=2.0.0
plotly>=5.18.0
praat-parselmouth>=0.4.3
pyaudio>=0.2.13